pyproject.toml
[project]
name = "Force-encryption"
version = "0.1.0"
description = "⭐ Fetch Star Wars data from swapi.info → Encrypt 🔒 → Log 📊"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Alex.D"}
]

dependencies = [
    "pandas>=2.0",
    "requests>=2.31",
    "cryptography>=43.0",
    "python-dotenv>=1.0",
    "colorlog>=6.9.0",
]

[tool.uv]
package = true